+++ client-app/shared/wishlists/components/wishlist-line-items.vue
@@ -46,10 +46,9 @@
 </template>
 
 <script setup lang="ts">
-import { ref } from "vue";
 import { ProductType } from "@/core/enums";
 import { InStock, CountInCart } from "@/shared/catalog";
-import type { ValidationErrorType } from "@/core/api/graphql/types";
+import { useLineItemsValidation } from "@/ui-kit";
 import type { PreparedLineItemType } from "@/core/types";
 
 interface IEmits {
@@ -65,7 +64,7 @@ interface IProp {
 const emit = defineEmits<IEmits>();
 defineProps<IProp>();
 
-const validationErrors = ref<ValidationErrorType[]>([]);
+const { validationErrors, setValidationStatus } = useLineItemsValidation();
 
 function addToCartDisabled(item: PreparedLineItemType) {
   return (
@@ -80,29 +79,5 @@ function changeCartItemQuantity(item: PreparedLineItemType, quantity: number): v
 
 function changeItemQuantity(item: PreparedLineItemType, quantity: number): void {
   emit("update:listItem", item, quantity);
-}
-
-function setValidationStatus(
-  item: PreparedLineItemType,
-  status: { isValid: true } | { isValid: false; errorMessage: string },
-): void {
-  const existingValidationError = validationErrors.value.find((error) => error.objectId === item.id);
-
-  if (!existingValidationError && !status.isValid) {
-    validationErrors.value.push({
-      objectId: item.id,
-      errorMessage: status.errorMessage,
-    });
-    return;
-  }
-
-  if (existingValidationError && !status.isValid) {
-    existingValidationError.errorMessage = status.errorMessage;
-    return;
-  }
-
-  if (existingValidationError && status.isValid) {
-    validationErrors.value = validationErrors.value.filter((error) => error.objectId !== item.id);
-  }
 }
 </script>
