+++ client-app/ui-kit/composables/index.ts
@@ -1,4 +1,5 @@
 export * from "./useAttrsOnly";
 export * from "./useComponentId";
+export * from "./useLineItemsValidation";
 export * from "./useListeners";
 export * from "./useQuantityValidationSchema";
